gpispace: scheduler: use a better matching degree which ensures that among the matching workers, the one with less capabilities is chosen

+++ sdpa/sdpa/daemon/scheduler/CoallocationScheduler.cpp
@@ -40,7 +40,7 @@ namespace sdpa
 
     namespace
     {
-      typedef std::tuple<double, int, unsigned long, double, worker_id_t> cost_deg_wid_t;
+      typedef std::tuple<double, double, unsigned long, double, worker_id_t> cost_deg_wid_t;
 
       struct min_cost_max_deg_comp
       {
